+++ CHANGELOG.md
@@ -7,9 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [1.8.0] - 2021-06-07
+## [1.10.0] - 2021-06-08
 
 - Dropdown Select component added
+
+## [1.9.0] - 2021-06-07
+
+- Tag component added
+- Enzyme serializer added
+
+## [1.8.0] - 2021-06-07
+
+- Dropdown component added
 
 ## [1.7.0] - 2021-06-04